+++ SensingKit/SKMotionActivity.m
@@ -1,5 +1,5 @@
 //
-//  SKActivity.m
+//  SKMotionActivity.m
 //  SensingKit
 //
 //  Copyright (c) 2014. Queen Mary University of London
@@ -22,22 +22,22 @@
 //  along with SensingKit-iOS.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-#import "SKActivity.h"
-#import "SKActivityData.h"
+#import "SKMotionActivity.h"
+#import "SKMotionActivityData.h"
 
 @import CoreMotion;
 
 
-@interface SKActivity ()
+@interface SKMotionActivity ()
 
 @property (nonatomic, strong) CMMotionActivityManager *motionActivityManager;
 
 @end
 
 
-@implementation SKActivity
+@implementation SKMotionActivity
 
-- (instancetype)initWithConfiguration:(SKActivityConfiguration *)configuration
+- (instancetype)initWithConfiguration:(SKMotionActivityConfiguration *)configuration
 {
     if (self = [super initWithConfiguration:configuration])
     {
@@ -52,9 +52,9 @@
 - (void)setConfiguration:(SKConfiguration *)configuration
 {
     // Check if the correct configuration type provided
-    if (configuration.class != SKActivityConfiguration.class)
+    if (configuration.class != SKMotionActivityConfiguration.class)
     {
-        NSLog(@"Wrong SKConfiguration class provided (%@) for sensor Activity.", configuration.class);
+        NSLog(@"Wrong SKConfiguration class provided (%@) for sensor Motion Activity.", configuration.class);
         abort();
     }
     
@@ -63,7 +63,7 @@
         super.configuration = configuration;
         
         // Cast the configuration instance
-        // SKActivityConfiguration *activityConfiguration = (SKActivityConfiguration *)configuration;
+        // SKMotionActivityConfiguration *motionActivityConfiguration = (SKMotionActivityConfiguration *)configuration;
         
         // Make the required updates on the sensor
         //
@@ -86,7 +86,7 @@
     {
         [self.motionActivityManager startActivityUpdatesToQueue:[NSOperationQueue currentQueue]
                                                     withHandler:^(CMMotionActivity *activity) {
-                                                        SKActivityData *data = [[SKActivityData alloc] initWithActivity:activity];
+                                                        SKMotionActivityData *data = [[SKMotionActivityData alloc] initWithMotionActivity:activity];
                                                         [self submitSensorData:data];
                                                     }];
     }